pyproject.toml
[tool.poetry]
name = "beebot"
version = "0.1.0"
description = "An AI agent that works"
authors = ["Erik Peterson <e@eriklp.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "beebot" }]

[tool.poetry.scripts]
beebot = 'beebot.initiator.cli:main'

[tool.poetry.dependencies]
# I hate this version specification but we can't use python-statemachine in 3.12
python = ">=3.10,<3.12"
langchain = ">=0.0.215"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
python-statemachine = "^2.1.0"
playwright = "^1.35.0"
lxml = "^4.9.3"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
google-api-python-client = "^2.93.0"
wolframalpha = "^5.0.0"
yoyo-migrations = "^8.2.0"
psycopg2-binary = { extras = ["ext"], version = "^2.9.6" }
beautifulsoup4 = "^4.12.2"
psutil = "^5.9.5"
coloredlogs = "^15.0.1"
uvicorn = { extras = ["standard"], version = "^0.23.1" }
fastapi = "^0.100.0"
wikipedia = "^1.4.0"
autopack-tools = "^0.4.0"
tortoise-orm = { extras = ["postgres"], version = "^0.19.3" }
asyncpg = "^0.28.0"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.276"
pytest = "^7.4.0"
colorama = "^0.4.6"
pytest-asyncio = "^0.21.1"
baserun = "^0.3.1"
agbenchmark = "^0.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
# Never enforce `E501` (line length violations). Newlines in strings sent to the LLM may impact its comprehension.
ignore = ["E501"]